pyproject.toml
[project]
name = "semantic-search"
version = "0.1.0"
description = "Десктопное приложение для семантического поиска по документам с использованием Doc2Vec"
authors = [
    {name = "Evgeny Odintsov",email = "ev1genial@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pymupdf (>=1.26.0,<2.0.0)",
    "pywin32 (>=310,<311); sys_platform == 'win32'",
    "python-docx (>=1.1.2,<2.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "gensim (>=4.3.3,<5.0.0)",
    "pyqt6 (>=6.9.1,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "click (>=8.2.1,<9.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]

[project.scripts]
semantic-search = "semantic_search.main:main"
semantic-search-cli = "semantic_search.main:cli_mode"

[tool.poetry]
packages = [{include = "semantic_search", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
pytest = "^8.4.0"
pytest-qt = "^4.4.0"
pyinstaller = "^6.14.1"
pytest-cov = "^6.2.1"
pytest-benchmark = "^5.1.0"
mypy = "^1.16.1"
pre-commit = "^4.2.0"
types-psutil = "^7.0.0.20250601"
types-pywin32 = "^310.0.0.20250516"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

